requirements.txt
Flask==2.0.3
gevent==21.12.0
h5py==3.6.0
numpy==1.22.3
Pillow==9.0.1
tensorflow-cpu
Werkzeug==2.0.3
torch==1.11.0
torchvision==0.12.0
keras
Keras-Preprocessing
opencv-python-headless
gunicorn
